requirements.txt
Django==5.0.3
mysqlclient==2.2.4
requests==2.31.0
django-cors-headers==4.3.1
sentry-sdk==1.39.2
psycopg2==2.9.9
sendgrid==6.11.0
tivars==0.8.2
python-dotenv==1.0.1
gunicorn==22.0.0
django-waffle==4.1.0
mysqlclient==2.2.4
django-storages[s3]==1.14.2
django-allauth==0.61.1
requests==2.31.0
requests-oauthlib==2.0.0
